requirement.txt
beautifulsoup4
rich
request